+++ reprozip/reprozip/filters.py
@@ -5,6 +5,7 @@
 from __future__ import division, print_function, unicode_literals
 
 import logging
+import re
 
 from reprozip.tracer.trace import TracedFile
 from reprozip.utils import irange, iteritems
@@ -13,13 +14,20 @@ from reprozip.utils import irange, iteritems
 logger = logging.getLogger('reprozip')
 
 
+_so_file = re.compile(br'\.so(\.[0-9]+)*$')
+
+
 def builtin(input_files, **kwargs):
     """Default heuristics for input files.
     """
     for i in irange(len(input_files)):
         lst = []
         for path in input_files[i]:
-            if path.unicodename[0] == '.' or path.ext in ('.pyc', '.so'):
+            if (
+                    # Hidden files
+                    path.unicodename[0] == '.' or
+                    # Shared libraries
+                    _so_file.search(path.name)):
                 logger.info("Removing input %s", path)
             else:
                 lst.append(path)
@@ -30,7 +38,7 @@ def builtin(input_files, **kwargs):
 def python(files, input_files, **kwargs):
     add = []
     for path, fi in iteritems(files):
-        if path.ext == '.pyc':
+        if path.ext == b'.pyc':
             pyfile = path.parent / path.stem + '.py'
             if pyfile.is_file():
                 if pyfile not in files:
@@ -43,7 +51,7 @@ def python(files, input_files, **kwargs):
     for i in irange(len(input_files)):
         lst = []
         for path in input_files[i]:
-            if path.ext in ('.py', '.pyc'):
+            if path.ext in (b'.py', b'.pyc'):
                 logger.info("Removing input %s", path)
             else:
                 lst.append(path)