+++ CHANGELOG.md
@@ -11,6 +11,24 @@ Features:
 * Configuration file contains the walltime taken by each run
 * It is now possible to upload or download any file via its full path
 
+1.0.16 (???)
+------------
+
+Bugfixes:
+* Fixed input/output file filter on Python 3 (to omit `.so`, `.pyc` etc files)
+
+Enhancements:
+* Use the [distro](https://distro.readthedocs.io/) module instead of the deprecated `platform.linux_distribution()` function to detect the distribution (the latter will be removed in Python 3.8).
+
+1.0.15 (2018-07-31)
+-------------------
+
+(reprounzip-qt only)
+
+Bugfixes:
+* Fixed running command from reprounzip-qt on Windows
+* Fixed using Jupyter from reprounzip-qt
+
 1.0.14 (2018-07-30)
 -------------------